git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@11215 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/domain/PropertyMapKey.java
@@ -39,7 +39,7 @@ public class PropertyMapKey implements Serializable, Comparable<PropertyMapKey>
     
     private Long qnameId;
     private Long localeId;
-    private Short listIndex;
+    private Integer listIndex;
     
     public PropertyMapKey()
     {
@@ -121,12 +121,12 @@ public class PropertyMapKey implements Serializable, Comparable<PropertyMapKey>
         this.localeId = localeId;
     }
 
-    public Short getListIndex()
+    public Integer getListIndex()
     {
         return listIndex;
     }
 
-    public void setListIndex(Short listIndex)
+    public void setListIndex(Integer listIndex)
     {
         this.listIndex = listIndex;
     }
